# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/platform/workspaces": {
      "get": {
        "operationId": "PlatformWorkspacesController_list",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 25,
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "trial",
                "active",
                "suspended",
                "closed"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid",
                            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                          },
                          "name": {
                            "type": "string"
                          },
                          "slug": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "trial",
                              "active",
                              "suspended",
                              "closed"
                            ]
                          },
                          "createdVia": {
                            "type": "string",
                            "enum": [
                              "platform",
                              "partner",
                              "self_service"
                            ]
                          },
                          "partnerWorkspaceId": {
                            "nullable": true,
                            "type": "string",
                            "format": "uuid",
                            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                          },
                          "trialEndsAt": {
                            "nullable": true,
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string"
                          },
                          "plan": {
                            "nullable": true,
                            "type": "string"
                          },
                          "modules": {
                            "type": "array",
                            "items": {
                              "type": "string",
                              "enum": [
                                "gap",
                                "materiality",
                                "actions",
                                "surveys",
                                "supply_chain",
                                "ranking",
                                "frameworks",
                                "ai",
                                "carbon"
                              ]
                            }
                          },
                          "limits": {
                            "type": "object",
                            "properties": {
                              "companies": {
                                "nullable": true,
                                "type": "integer",
                                "minimum": 0,
                                "maximum": 9007199254740991
                              },
                              "users": {
                                "nullable": true,
                                "type": "integer",
                                "minimum": 0,
                                "maximum": 9007199254740991
                              },
                              "projectsPerYear": {
                                "nullable": true,
                                "type": "integer",
                                "minimum": 0,
                                "maximum": 9007199254740991
                              },
                              "clientWorkspaces": {
                                "nullable": true,
                                "type": "integer",
                                "minimum": 0,
                                "maximum": 9007199254740991
                              },
                              "aiMonthlyUsd": {
                                "nullable": true,
                                "type": "integer",
                                "minimum": 0,
                                "maximum": 9007199254740991
                              }
                            },
                            "additionalProperties": false
                          },
                          "companies": {
                            "type": "integer",
                            "minimum": -9007199254740991,
                            "maximum": 9007199254740991
                          },
                          "members": {
                            "type": "integer",
                            "minimum": -9007199254740991,
                            "maximum": 9007199254740991
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "slug",
                          "status",
                          "createdVia",
                          "partnerWorkspaceId",
                          "trialEndsAt",
                          "createdAt",
                          "plan",
                          "modules",
                          "limits",
                          "companies",
                          "members"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "nextCursor": {
                      "nullable": true,
                      "type": "string"
                    }
                  },
                  "required": [
                    "items",
                    "nextCursor"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "platform"
        ]
      }
    },
    "/v1/platform/workspaces/{id}": {
      "patch": {
        "operationId": "PlatformWorkspacesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "trial",
                      "active",
                      "suspended",
                      "closed"
                    ]
                  },
                  "trialEndsAt": {
                    "nullable": true,
                    "type": "string",
                    "format": "date-time",
                    "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "name": {
                      "type": "string"
                    },
                    "slug": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "trial",
                        "active",
                        "suspended",
                        "closed"
                      ]
                    },
                    "createdVia": {
                      "type": "string",
                      "enum": [
                        "platform",
                        "partner",
                        "self_service"
                      ]
                    },
                    "partnerWorkspaceId": {
                      "nullable": true,
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "trialEndsAt": {
                      "nullable": true,
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "plan": {
                      "nullable": true,
                      "type": "string"
                    },
                    "modules": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "enum": [
                          "gap",
                          "materiality",
                          "actions",
                          "surveys",
                          "supply_chain",
                          "ranking",
                          "frameworks",
                          "ai",
                          "carbon"
                        ]
                      }
                    },
                    "limits": {
                      "type": "object",
                      "properties": {
                        "companies": {
                          "nullable": true,
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 9007199254740991
                        },
                        "users": {
                          "nullable": true,
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 9007199254740991
                        },
                        "projectsPerYear": {
                          "nullable": true,
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 9007199254740991
                        },
                        "clientWorkspaces": {
                          "nullable": true,
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 9007199254740991
                        },
                        "aiMonthlyUsd": {
                          "nullable": true,
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 9007199254740991
                        }
                      },
                      "additionalProperties": false
                    },
                    "companies": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "members": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "slug",
                    "status",
                    "createdVia",
                    "partnerWorkspaceId",
                    "trialEndsAt",
                    "createdAt",
                    "plan",
                    "modules",
                    "limits",
                    "companies",
                    "members"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "platform"
        ]
      }
    },
    "/v1/platform/workspaces/{id}/entitlements": {
      "put": {
        "operationId": "PlatformWorkspacesController_putEntitlements",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "plan": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 64
                  },
                  "modules": {},
                  "limits": {
                    "type": "object",
                    "properties": {
                      "companies": {
                        "nullable": true,
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 9007199254740991
                      },
                      "users": {
                        "nullable": true,
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 9007199254740991
                      },
                      "projectsPerYear": {
                        "nullable": true,
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 9007199254740991
                      },
                      "clientWorkspaces": {
                        "nullable": true,
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 9007199254740991
                      },
                      "aiMonthlyUsd": {
                        "nullable": true,
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 9007199254740991
                      }
                    },
                    "additionalProperties": false
                  }
                },
                "required": [
                  "plan",
                  "modules",
                  "limits"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "plan": {
                      "type": "string"
                    },
                    "modules": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "enum": [
                          "gap",
                          "materiality",
                          "actions",
                          "surveys",
                          "supply_chain",
                          "ranking",
                          "frameworks",
                          "ai",
                          "carbon"
                        ]
                      }
                    },
                    "limits": {
                      "type": "object",
                      "properties": {
                        "companies": {
                          "nullable": true,
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 9007199254740991
                        },
                        "users": {
                          "nullable": true,
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 9007199254740991
                        },
                        "projectsPerYear": {
                          "nullable": true,
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 9007199254740991
                        },
                        "clientWorkspaces": {
                          "nullable": true,
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 9007199254740991
                        },
                        "aiMonthlyUsd": {
                          "nullable": true,
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 9007199254740991
                        }
                      },
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "plan",
                    "modules",
                    "limits"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "platform"
        ]
      }
    }
  },
  "info": {
    "title": "ResiliSense API",
    "description": "ResiliSense 2.0 REST API. Specs: docs/revamp/modules. Errors are RFC 9457 problem+json.",
    "version": "2.0.0-alpha",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}